fix(#98): address Aikido review on PR #98
Aikido flagged one logic bug and three code-quality issues on this PR.

Logic bug (gateway/src/api.ts) — the multi-mount IDE project root fell back
with `commonParentPath(...) || '/workspaces'`, but commonParentPath() returns
'/' (truthy) when the mounts share no leading segment, so the documented
'/workspaces' default was unreachable and such containers opened at '/'. The
helper plus the fallback now live in gateway/src/workspace-root.ts — pure path
arithmetic, no Fastify/DB/Docker to load — and workspace-root.test.ts pins both
the shared-parent case and the fallback.

Duplicated validation (gateway/src/db.ts) — validateGroupKeys repeated
validateFolderMappingKeys verbatim (allowlist, unknown-key detection, throw,
filtered return). Both now delegate to one generic validateUpdateKeys(), so the
SQL-identifier allowlist that closes finding #9 has a single implementation.

Single responsibility + function length (gateway/src/firewall-groups.ts) — the
module mixed envelope validation, DB persistence and folder file I/O, and
syncGroupsToFolder did discovery, filename selection, writing, DB retagging and
pruning inline. Split by responsibility:
  firewall-group-envelope.ts  envelope shape + fail-closed validation (pure)
  firewall-group-store.ts     DB reads/writes for groups and their rules
  firewall-rules-folder.ts    the team folder on disk (reload + sync)
firewall-groups.ts stays as the public surface (re-exports only), so api.ts and
the existing tests are untouched. Both long functions are now composed of named
helpers (pickGroupFile, writeGroupFile, pruneOrphanEnvelopes, parseEnvelopeFiles,
applyParsedEnvelopes), and the member-rules query that export and apply share is
one memberRulesSql() so the two can no longer disagree on a group's contents.

Behaviour-preserving throughout: gateway suite green (364 tests, 23 files),
including the firewall-groups cases covering merge/replace import, apply,
folder reload and folder sync.

The fourth Aikido comment (cli/src/migrate.ts parseYaml) is about code already
merged into main — after the rebase that file is no longer part of this PR's
diff.
